refactor: connect four environment  version 2
